Move local `EventStore` impl. to the package level
